Move some files around to make the file structure the same across sites and techniques (just for easier parsing of the file tree)
