+++ app/src/features/surveys/sampling-information/SamplingRouter.tsx
@@ -4,8 +4,8 @@ import { DialogContextProvider } from 'contexts/dialogContext';
 import { SamplingSiteManagePage } from 'features/surveys/sampling-information/manage/SamplingSiteManagePage';
 import { CreateSamplingSitePage } from 'features/surveys/sampling-information/sites/create/CreateSamplingSitePage';
 import { EditSamplingSitePage } from 'features/surveys/sampling-information/sites/edit/EditSamplingSitePage';
-import { CreateTechniquePage } from 'features/surveys/sampling-information/techniques/form/create/CreateTechniquePage';
-import { EditTechniquePage } from 'features/surveys/sampling-information/techniques/form/edit/EditTechniquePage';
+import { CreateTechniquePage } from 'features/surveys/sampling-information/techniques/create/CreateTechniquePage';
+import { EditTechniquePage } from 'features/surveys/sampling-information/techniques/edit/EditTechniquePage';
 import { Switch } from 'react-router';
 import RouteWithTitle from 'utils/RouteWithTitle';
 import { getTitle } from 'utils/Utils';
