+++ app/src/features/surveys/sampling-information/techniques/components/attributes/TechniqueAttributesForm.tsx
@@ -7,7 +7,7 @@ import {
   CreateTechniqueFormValues,
   TechniqueAttributeFormValues,
   UpdateTechniqueFormValues
-} from 'features/surveys/sampling-information/techniques/form/components/TechniqueFormContainer';
+} from 'features/surveys/sampling-information/techniques/components/TechniqueFormContainer';
 import { FieldArray, FieldArrayRenderProps, useFormikContext } from 'formik';
 import { ITechniqueAttributeQualitative, ITechniqueAttributeQuantitative } from 'interfaces/useReferenceApi.interface';
 import { TransitionGroup } from 'react-transition-group';
