+++ app/src/features/surveys/sampling-information/techniques/components/attributes/components/utils.ts
@@ -1,5 +1,5 @@
 import { IAutocompleteFieldOption } from 'components/fields/AutocompleteField';
-import { TechniqueAttributeFormValues } from 'features/surveys/sampling-information/techniques/form/components/TechniqueFormContainer';
+import { TechniqueAttributeFormValues } from 'features/surveys/sampling-information/techniques/components/TechniqueFormContainer';
 import { ITechniqueAttributeQualitative, ITechniqueAttributeQuantitative } from 'interfaces/useReferenceApi.interface';
 
 /**
