+++ app/src/features/surveys/sampling-information/techniques/components/details/TechniqueDetailsForm.tsx
