+++ app/src/features/surveys/sampling-information/techniques/components/general-information/TechniqueGeneralInformationForm.tsx
@@ -6,7 +6,7 @@ import { ISelectWithSubtextFieldOption } from 'components/fields/SelectWithSubte
 import {
   CreateTechniqueFormValues,
   UpdateTechniqueFormValues
-} from 'features/surveys/sampling-information/techniques/form/components/TechniqueFormContainer';
+} from 'features/surveys/sampling-information/techniques/components/TechniqueFormContainer';
 
 import { useFormikContext } from 'formik';
 import { useCodesContext } from 'hooks/useContext';
