+++ app/src/features/surveys/sampling-information/techniques/edit/EditTechniquePage.tsx
@@ -8,37 +8,33 @@ import Paper from '@mui/material/Paper';
 import Stack from '@mui/material/Stack';
 import Typography from '@mui/material/Typography';
 import PageHeader from 'components/layout/PageHeader';
-import { CreateTechniqueI18N } from 'constants/i18n';
+import { EditTechniqueI18N } from 'constants/i18n';
 import TechniqueFormContainer, {
-  CreateTechniqueFormValues
-} from 'features/surveys/sampling-information/techniques/form/components/TechniqueFormContainer';
+  UpdateTechniqueFormValues
+} from 'features/surveys/sampling-information/techniques/components/TechniqueFormContainer';
 import { FormikProps } from 'formik';
 import { APIError } from 'hooks/api/useAxios';
 import { useBiohubApi } from 'hooks/useBioHubApi';
 import { useDialogContext, useProjectContext, useSurveyContext } from 'hooks/useContext';
+import useDataLoader from 'hooks/useDataLoader';
 import { SKIP_CONFIRMATION_DIALOG, useUnsavedChangesDialog } from 'hooks/useUnsavedChangesDialog';
-import { ICreateTechniqueRequest } from 'interfaces/useTechniqueApi.interface';
-import { useRef, useState } from 'react';
-import { Prompt, useHistory } from 'react-router';
+import { IUpdateTechniqueRequest } from 'interfaces/useTechniqueApi.interface';
+import { useEffect, useRef, useState } from 'react';
+import { Prompt, useHistory, useParams } from 'react-router';
 import { Link as RouterLink } from 'react-router-dom';
-
-const initialTechniqueFormValues: CreateTechniqueFormValues = {
-  name: '',
-  description: '',
-  distance_threshold: null,
-  method_lookup_id: null,
-  attractants: [],
-  attributes: []
-};
+import { v4 } from 'uuid';
 
 /**
- * Renders the body content of the create technique page.
+ * Renders the body content of the Technique page.
  *
  * @return {*}
  */
-export const CreateTechniquePage = () => {
+export const EditTechniquePage = () => {
   const history = useHistory();
   const biohubApi = useBiohubApi();
+
+  const urlParams: Record<string, string | number | undefined> = useParams();
+  const methodTechniqueId = Number(urlParams['method_technique_id']);
 
   const surveyContext = useSurveyContext();
   const projectContext = useProjectContext();
@@ -48,41 +44,83 @@ export const CreateTechniquePage = () => {
 
   const [isSubmitting, setIsSubmitting] = useState(false);
 
-  const formikRef = useRef<FormikProps<CreateTechniqueFormValues>>(null);
+  const formikRef = useRef<FormikProps<UpdateTechniqueFormValues>>(null);
+
+  const techniqueDataLoader = useDataLoader(() =>
+    biohubApi.technique.getTechniqueById(surveyContext.projectId, surveyContext.surveyId, methodTechniqueId)
+  );
+
+  useEffect(() => {
+    techniqueDataLoader.load();
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
 
   if (!surveyContext.surveyDataLoader.data || !projectContext.projectDataLoader.data) {
     return <CircularProgress className="pageProgress" size={40} />;
   }
 
-  const handleSubmit = async (values: CreateTechniqueFormValues) => {
+  const technique = techniqueDataLoader.data;
+
+  if (!technique) {
+    return <CircularProgress className="pageProgress" size={40} />;
+  }
+
+  const initialTechniqueValues: UpdateTechniqueFormValues = {
+    method_technique_id: technique.method_technique_id,
+    name: technique.name ?? null,
+    description: technique?.description ?? null,
+    distance_threshold: technique?.distance_threshold ?? null,
+    method_lookup_id: technique?.method_lookup_id ?? null,
+    attractants: technique?.attractants,
+    attributes: [
+      ...(technique?.attributes.qualitative_attributes.map((attribute) => ({
+        _id: v4(), // A temporary unique id for react keys, etc, as the attribute_id is not unique
+        attribute_id: attribute.method_technique_attribute_qualitative_id ?? null,
+        attribute_lookup_id: attribute.method_lookup_attribute_qualitative_id,
+        attribute_value: attribute.method_lookup_attribute_qualitative_option_id,
+        attribute_type: 'qualitative' as const
+      })) ?? []),
+      ...(technique?.attributes.quantitative_attributes.map((attribute) => ({
+        _id: v4(), // A temporary unique id for react keys, etc, as the attribute_id is not unique
+        attribute_id: attribute.method_technique_attribute_quantitative_id ?? null,
+        attribute_lookup_id: attribute.method_lookup_attribute_quantitative_id,
+        attribute_value: attribute.value,
+        attribute_type: 'quantitative' as const
+      })) ?? [])
+    ]
+  };
+
+  const handleSubmit = async (values: UpdateTechniqueFormValues) => {
     try {
       setIsSubmitting(true);
 
-      // Parse the form data into the request format
-      const createTechniqueRequestData: ICreateTechniqueRequest = {
+      const formattedTechniqueObject: IUpdateTechniqueRequest = {
         ...values,
         attributes: {
-          qualitative_attributes: values.attributes
-            .filter(({ attribute_type }) => attribute_type === 'qualitative')
-            .map((item) => ({
-              method_technique_attribute_qualitative_id: null,
-              method_lookup_attribute_qualitative_id: item.attribute_lookup_id,
-              method_lookup_attribute_qualitative_option_id: item.attribute_value as string
-            })),
           quantitative_attributes: values.attributes
-            .filter(({ attribute_type }) => attribute_type === 'quantitative')
-            .map((item) => ({
-              method_technique_attribute_quantitative_id: null,
-              method_lookup_attribute_quantitative_id: item.attribute_lookup_id,
-              value: item.attribute_value as number
+            .filter((attribute) => attribute.attribute_type === 'quantitative')
+            .map((attribute) => ({
+              method_technique_attribute_quantitative_id: attribute.attribute_id,
+              method_lookup_attribute_quantitative_id: attribute.attribute_lookup_id,
+              value: attribute.attribute_value as number
+            })),
+          qualitative_attributes: values.attributes
+            .filter((attribute) => attribute.attribute_type === 'qualitative')
+            .map((attribute) => ({
+              method_technique_attribute_qualitative_id: attribute.attribute_id,
+              method_lookup_attribute_qualitative_id: attribute.attribute_lookup_id,
+              method_lookup_attribute_qualitative_option_id: attribute.attribute_value as string
             }))
         }
       };
 
-      // Create the technique
-      await biohubApi.technique.createTechniques(surveyContext.projectId, surveyContext.surveyId, [
-        createTechniqueRequestData
-      ]);
+      // Update the technique
+      await biohubApi.technique.updateTechnique(
+        surveyContext.projectId,
+        surveyContext.surveyId,
+        methodTechniqueId,
+        formattedTechniqueObject
+      );
 
       // Refresh the context, so the next page loads with the latest data
       surveyContext.techniqueDataLoader.refresh(surveyContext.projectId, surveyContext.surveyId);
@@ -95,8 +133,8 @@ export const CreateTechniquePage = () => {
     } catch (error) {
       setIsSubmitting(false);
       dialogContext.setErrorDialog({
-        dialogTitle: CreateTechniqueI18N.createErrorTitle,
-        dialogText: CreateTechniqueI18N.createErrorText,
+        dialogTitle: EditTechniqueI18N.createErrorTitle,
+        dialogText: EditTechniqueI18N.createErrorText,
         dialogError: (error as APIError).message,
         dialogErrorDetails: (error as APIError)?.errors,
         onClose: () => {
@@ -114,7 +152,7 @@ export const CreateTechniquePage = () => {
     <>
       <Prompt when={true} message={locationChangeInterceptor} />
       <PageHeader
-        title="Create New Technique"
+        title="Edit Technique"
         breadCrumbJSX={
           <Breadcrumbs
             aria-label="breadcrumb"
@@ -141,7 +179,7 @@ export const CreateTechniquePage = () => {
               Manage Sampling Information
             </Link>
             <Typography component="span" variant="body2" color="textSecondary">
-              Create New Technique
+              Edit Technique
             </Typography>
           </Breadcrumbs>
         }
@@ -170,7 +208,7 @@ export const CreateTechniquePage = () => {
       <Container maxWidth="xl" sx={{ py: 3 }}>
         <Paper sx={{ p: 5 }}>
           <TechniqueFormContainer
-            initialData={initialTechniqueFormValues}
+            initialData={initialTechniqueValues}
             handleSubmit={(formikData) => handleSubmit(formikData)}
             formikRef={formikRef}
           />
