+++ app/src/features/surveys/sampling-information/techniques/table/SamplingTechniqueTable.tsx
