+++ app/src/features/surveys/view/components/sampling-data/components/SurveySamplingTabs.tsx
@@ -7,7 +7,7 @@ import ToggleButton from '@mui/material/ToggleButton';
 import ToggleButtonGroup from '@mui/material/ToggleButtonGroup';
 import { LoadingGuard } from 'components/loading/LoadingGuard';
 import { SkeletonTable } from 'components/loading/SkeletonLoaders';
-import { SurveyPeriodsTable } from 'features/surveys/sampling-information/sites/manage/table/period/SurveyPeriodsTable';
+import { SurveyPeriodsTable } from 'features/surveys/sampling-information/sites/table/period/SurveyPeriodsTable';
 import { SurveySitesTable } from 'features/surveys/view/components/sampling-data/components/SurveySitesTable';
 import { SurveyTechniquesTable } from 'features/surveys/view/components/sampling-data/components/SurveyTechniquesTable';
 import { useSurveyContext } from 'hooks/useContext';
